vault-*.sh: reference all script locations using _VAULT_SCRIPT_DIR
This allows one to set the location of the setup scripts in a different location than the config/state files. I.e. in the snap we will have the setup scripts inside $SNAP which is read-only, and the config files and such will live in $SNAP_DATA, which is writable.
If the variable isn't defined it defaults to /vault, so there should be no difference for the docker setup which doesn't set any environment variables for where to find things.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>